refactor:fuzz: consolidate fuzz types
Add common types to the lib
