+++ fuzz/fuzz_targets/select_coins_srd.rs
@@ -1,27 +1,11 @@
 #![no_main]
 
-use arbitrary::{Arbitrary, Unstructured, Result};
-use bitcoin_units::{FeeRate, Amount, Weight};
-use bitcoin_coin_selection::{select_coins_srd, WeightedUtxo};
+use arbitrary::{Arbitrary, Unstructured};
+use bitcoin_coin_selection::select_coins_srd;
+use bitcoin_coin_selection_fuzz::UtxoPool;
+use bitcoin_units::{Amount, Weight};
 use libfuzzer_sys::fuzz_target;
 use rand::thread_rng;
-
-#[derive(Debug)]
-pub struct UtxoPool {
-    pub utxos: Vec<WeightedUtxo>,
-}
-
-impl<'a> Arbitrary<'a> for UtxoPool {
-    fn arbitrary(u: &mut Unstructured<'a>) -> Result<Self> {
-        let init: Vec<(Amount, Weight)> = Vec::arbitrary(u)?;
-        let fee_rate = FeeRate::arbitrary(u).unwrap();
-        let lt_fee_rate = FeeRate::arbitrary(u).unwrap();
-        let pool: Vec<WeightedUtxo> =
-            init.iter().filter_map(|i| WeightedUtxo::new(i.0, i.1, fee_rate, lt_fee_rate)).collect();
-
-        Ok(UtxoPool { utxos: pool })
-    }
-}
 
 fuzz_target!(|data: &[u8]| {
     let mut u = Unstructured::new(&data);
